Implement basic network client
